requirements.txt
beautifulsoup4==4.12.2
Flask==3.0.0
Flask-SocketIO==5.3.6
passlib==1.7.4
psutil==5.9.5
pyperclip==1.8.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.0